[Client/ServerWindow] Fix continuously opening ErrorPopup
After receiving an error while joining the whole window would be considered part of the listbox, therefore opening the same popup on every click.
Most likely this happened because the event was not properly handled.
By changing to a ItemsControl and Buttons instead of a ListBox the listbox-specific hacks (e.g. setting the selected item to null inside the handler) can be avoided and all events are properly handled