Make Location a normal enum
It does not make sense for a card to be in more than one location at once.
The only reason location was a bitfield in the first place was
the ability to filter if a card is in one of many locations.
This was never used outside of Location.ALL anyways and can still be achieved by
the a slightly more verbose chain of checks if one so desired.