Ditch JSON in favor of brief
While the reasons to ditch JSON (e.g. inability to send binary data, being a text-based format) might not really hold up under scrutiny and are driven mostly by a feeling of a JSON-based infrastructure just not being as elegant as it could be, it was a quite cool experience to experiment with different existing formats like ProtoBuf, FlatBuf, Cap'n'Proto and thrift, and writing my own format brief in the end.

There shouldn't be that many observable changes apart from Deck Edit opening quite a bit faster.