[Client]: Fix networking hogging RAM and CPU like crazy
Previously, the backround tasks listening for incoming packets would
busy loop until they received a packet. This was done because with the
previous architecture receiving could not be blocking, as all threads
were written with the assumption that they were the sole owner of the
shared NetworkStream. Since this was of course not the case, it occured
quite often that the background thread snatched a packet from the main
thread, effectively deadlocking it waiting for a packet that was already
gone. Previously, this was solved by liberally spraying mutexes all
around the codebase where it seemed appropriate and not reading from the
stream non-blockingly in the background thread, since with just a mutex
the background thread could potentially hog the mutex waiting for a
packet while another thread wants to send, leading to a deadlock because
the server expects to receive a packet.
This new solution instead only lets the background thread listen for
packets, shoving them into a thread-safe, blocking queue if they could
not be handled by that thread directly, hopefully avoiding the need for
a mutex while still preserving the straight-forward nature of the code
just writing followed by reading, just not from the stream but from the
queue now.